Move middleware pkg to web and standardize responses
